# use context managers for database cursor objects & stream results