… chia/full_node/coin_store.py