To make sure this doesn't cause any performance regressions, I ran both the block_store and coin_store benchmarks.
I would have expected this patch to not be any worse, but maybe a little bit faster. using `async for` to stream the sqlite results back allows pipelining reading from disk and parsing.

Are these timings noise? Is the code change worth it anyway?

| test | before | after | after / before
---|---|---|---
| block_store v1 | 333.8625s | 339.1758s | 101.6 %
| block_store v2 | 306.5919s | 319.9539s | 104.4 %
| coin_store v1 | 228.9748s | 219.8814s | 96.0 %
| coin_store v2 | 193.9898s | 195.0622s | 100.6 %

## This patch

block_store
```
version 1
89.2887s, add_full_block
52.8452s, get_full_block
5.8553s, get_full_block_bytes
53.1792s, get_full_blocks_at
7.2331s, get_block_records_by_hash
54.2095s, get_blocks_by_hash
5.4857s, get_block_record
0.6965s, get_block_records_in_range
0.0057s, get_block_records_close_to_peak
4.2023s, get_block_record
66.1747s, get_random_not_compactified
all tests completed in 339.1758s
database size: 2418.156 MB
version 2
44.2622s, add_full_block
52.6760s, get_full_block
5.0684s, get_full_block_bytes
53.2927s, get_full_blocks_at
7.9133s, get_block_records_by_hash
53.9895s, get_blocks_by_hash
6.5078s, get_block_record
0.7906s, get_block_records_in_range
0.0066s, get_block_records_close_to_peak
4.4581s, get_block_record
90.9888s, get_random_not_compactified
all tests completed in 319.9539s
database size: 357.474 MB
```

coin_store
```
version 1
Building database 
49.3254s, MOSTLY ADDITIONS additions: 400400.0 removals: 20000.0
8.0602s, MOSTLY REMOVALS additions: 600 removals: 140000
67.6810s, FULLBLOCKS additions: 400400 removals: 400000
1.4604s, GET RECORDS BY NAMES with spent 200 lookups found 40000 coins in total
66.7356s, GET RECORDS BY NAMES without spent 200 lookups found 20641 coins in total
26.6187s, GET COINS REMOVED AT HEIGHT 800 blocks, found 580000 coins in total
all tests completed in 219.8814s
database size: 638.099 MB
version 2
Building database 
44.3007s, MOSTLY ADDITIONS additions: 400400.0 removals: 20000.0
7.6507s, MOSTLY REMOVALS additions: 600 removals: 140000
64.7445s, FULLBLOCKS additions: 400400 removals: 400000
2.3373s, GET RECORDS BY NAMES with spent 200 lookups found 40000 coins in total
51.2288s, GET RECORDS BY NAMES without spent 200 lookups found 20693 coins in total
24.8002s, GET COINS REMOVED AT HEIGHT 800 blocks, found 580000 coins in total
all tests completed in 195.0622s
database size: 353.669 MB
```

## Prior commit:

block_store:
```
version 1
91.3370s, add_full_block
52.0561s, get_full_block
5.4880s, get_full_block_bytes
50.7568s, get_full_blocks_at
5.7596s, get_block_records_by_hash
53.2185s, get_blocks_by_hash
7.0929s, get_block_record
0.6808s, get_block_records_in_range
0.0057s, get_block_records_close_to_peak
4.8409s, get_block_record
62.6261s, get_random_not_compactified
all tests completed in 333.8625s
database size: 2418.156 MB
version 2
45.1364s, add_full_block
51.2920s, get_full_block
5.7642s, get_full_block_bytes
49.4884s, get_full_blocks_at
6.8339s, get_block_records_by_hash
52.1692s, get_blocks_by_hash
6.5498s, get_block_record
0.6922s, get_block_records_in_range
0.0062s, get_block_records_close_to_peak
4.2534s, get_block_record
84.4062s, get_random_not_compactified
all tests completed in 306.5919s
database size: 357.474 MB
```

coin_store
```
version 1
Building database 
56.8906s, MOSTLY ADDITIONS additions: 400400.0 removals: 20000.0
8.2210s, MOSTLY REMOVALS additions: 600 removals: 140000
70.9838s, FULLBLOCKS additions: 400400 removals: 400000
1.3132s, GET RECORDS BY NAMES with spent 200 lookups found 40000 coins in total
67.1411s, GET RECORDS BY NAMES without spent 200 lookups found 20641 coins in total
24.4251s, GET COINS REMOVED AT HEIGHT 800 blocks, found 580000 coins in total
all tests completed in 228.9748s
database size: 638.091 MB
version 2
Building database 
44.5126s, MOSTLY ADDITIONS additions: 400400.0 removals: 20000.0
7.9520s, MOSTLY REMOVALS additions: 600 removals: 140000
63.4308s, FULLBLOCKS additions: 400400 removals: 400000
2.1089s, GET RECORDS BY NAMES with spent 200 lookups found 40000 coins in total
49.9227s, GET RECORDS BY NAMES without spent 200 lookups found 20693 coins in total
26.0627s, GET COINS REMOVED AT HEIGHT 800 blocks, found 580000 coins in total
all tests completed in 193.9898s
database size: 353.620 MB
```