@arvidn 2021-12-16
I'm having second thoughts about using `async for` to stream results back. I suspect that might contributing to slowing this down. It might add more context switches for instance.