@arvidn 2021-12-16
@altendky I dropped all the `async for` and incorporated your suggested changes. The new timings are closer (but I just ran the new ones once).

This commit:

block_store:
```
version 1
89.6409s, add_full_block
53.2567s, get_full_block
5.2159s, get_full_block_bytes
51.1661s, get_full_blocks_at
5.8824s, get_block_records_by_hash
52.6979s, get_blocks_by_hash
7.1451s, get_block_record
0.6540s, get_block_records_in_range
0.0053s, get_block_records_close_to_peak
4.6177s, get_block_record
66.4469s, get_random_not_compactified
all tests completed in 336.7288s
database size: 2418.156 MB
version 2
44.0412s, add_full_block
52.3799s, get_full_block
5.0736s, get_full_block_bytes
50.2342s, get_full_blocks_at
6.3605s, get_block_records_by_hash
53.0350s, get_blocks_by_hash
6.6490s, get_block_record
0.7128s, get_block_records_in_range
0.0061s, get_block_records_close_to_peak
4.5579s, get_block_record
88.9816s, get_random_not_compactified
all tests completed in 312.0317s
database size: 357.474 MB
```

coin_store:
```
version 1
Building database
52.9241s, MOSTLY ADDITIONS additions: 400400.0 removals: 20000.0
7.6849s, MOSTLY REMOVALS additions: 600 removals: 140000
72.3118s, FULLBLOCKS additions: 400400 removals: 400000
1.4111s, GET RECORDS BY NAMES with spent 200 lookups found 40000 coins in total
68.3960s, GET RECORDS BY NAMES without spent 200 lookups found 20641 coins in total
24.7959s, GET COINS REMOVED AT HEIGHT 800 blocks, found 580000 coins in total
all tests completed in 227.5239s
database size: 638.095 MB
version 2
Building database
45.5893s, MOSTLY ADDITIONS additions: 400400.0 removals: 20000.0
7.9858s, MOSTLY REMOVALS additions: 600 removals: 140000
64.8043s, FULLBLOCKS additions: 400400 removals: 400000
2.1198s, GET RECORDS BY NAMES with spent 200 lookups found 40000 coins in total
50.4896s, GET RECORDS BY NAMES without spent 200 lookups found 20693 coins in total
23.2696s, GET COINS REMOVED AT HEIGHT 800 blocks, found 580000 coins in total
all tests completed in 194.2584s
database size: 353.620 MB
```