@arvidn 2021-12-20
Although that's probably a good rule of thumb, when it comes to async processes, I don't think it's necessarily always true. For instance, perhaps the exiting of the context will synchronize (and wait for) some background task completing. If you can defer this synchronization, you may increase concurrency.

But perhaps there aren't any examples of that in the real world.